chore(types): mypy frontier_ops/ to 0 errors with a 4-module tracked baseline
Add mypy to [dev] with a lenient config (ignore_missing_imports, untyped
defs unchecked). Fix ~50 real annotations across the library: Optional
defaults, missing var annotations, max(key=dict.get) overload traps, a
Boundary loop-variable shadowing bug in constitution.py's cross-term
metric assembly (rename only, math identical), lowercase 'any' return
type, and Optional-attribute EWMA pairs (narrow type: ignore comments in
newma/trend, whose logic is off-limits per docs/LEGACY.md).

Tracked baseline (documented in pyproject): sensing.modes (quarantined)
and three ATBench glue modules keep ignore_errors until next touched.
CI lint job gains a non-blocking mypy step.

+++ frontier_ops/sensing/efference.py
@@ -217,7 +217,7 @@ class EfferenceCopyPredictor:
 
         if self.error_history:
             historical_mean = np.mean(self.error_history[-20:])
-            surprise_ratio = weighted_magnitude / (historical_mean + 1e-8)
+            surprise_ratio = float(weighted_magnitude / (historical_mean + 1e-8))
         else:
             surprise_ratio = 1.0
 
@@ -320,7 +320,7 @@ class EfferenceCopyPredictor:
         alpha = self.ewma_alpha
         diff = actual - self._ewma_mean  # Compute diff BEFORE updating mean
         self._ewma_mean = alpha * actual + (1 - alpha) * self._ewma_mean
-        self._ewma_var = alpha * (diff ** 2) + (1 - alpha) * self._ewma_var
+        self._ewma_var = alpha * (diff ** 2) + (1 - alpha) * self._ewma_var  # type: ignore[operator]  # set with _ewma_mean
 
     # ─── Tier 3: Kalman Filter ────────────────────────────────────────
